externalize server-side collection functions to services
