+++ app/services/booster.ts
@@ -6,7 +6,7 @@ import {
   PkmAltForms,
   PkmAltFormsByPkm
 } from "../config"
-import { CollectionUtils, createBooster } from "../core/collection"
+import { CollectionUtils } from "../core/collection"
 import UserMetadata from "../models/mongo-models/user-metadata"
 import { getPokemonData } from "../models/precomputed/precomputed-pokemon-data"
 import { CollectionEmotions, Emotion, PkmWithCustom, Title } from "../types"
@@ -22,6 +22,7 @@ import {
   IPokemonCollectionItemMongo,
   IUserMetadataMongo
 } from "../types/interfaces/UserMetadata"
+import { createBooster } from "./collection"
 
 export type OpenBoosterResult = {
   userDoc: IUserMetadataMongo
